pyproject.toml
[project]
name = "yonk-code-robomonkey"
version = "0.1.0"
description = "Local-first RoboMonkey MCP server with Postgres+pgvector+FTS+tags+docs"
requires-python = ">=3.11"
dependencies = [
  "asyncpg>=0.29.0",
  "pydantic>=2.7.0",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "rich>=13.7.1",
  "watchdog>=4.0.1",
  "pathspec>=0.12.1",
  "tree-sitter==0.20.4",
  "tree-sitter-languages>=1.10.2",
  "typing-extensions>=4.11.0",
  "orjson>=3.10.0",
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "pyyaml>=6.0.0",
]

[project.scripts]
robomonkey = "yonk_code_robomonkey.cli.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
